ingress/requirements.txt
cachetools==3.0.0
certifi==2018.10.15
chardet==3.0.4
google-api-core==1.5.2
google-auth==1.6.1
google-cloud-core==0.28.1
google-cloud-error-reporting==0.30.0
google-cloud-logging==1.8.0
google-cloud-pubsub==0.38.0
googleapis-common-protos==1.5.5
grpc-google-iam-v1==0.11.4
grpcio==1.16.1
idna==2.7
jmespath==0.9.3
marshmallow==3.0.0b16
protobuf==3.6.1
pyasn1==0.4.4
pyasn1-modules==0.2.2
pytz==2018.7
PyYAML==3.13
requests==2.20.1
rsa==4.0
six==1.11.0
urllib3==1.24.1
